requirements.txt
# Core Framework
flask==2.3.3
flask-socketio==5.3.4
flask-cors==4.0.0
flask-jwt-extended==4.5.2
python-socketio==5.9.0
eventlet==0.33.3
requests==2.31.0

# External Device Monitoring
pyudev==0.24.0; platform_system == "Linux"
pywin32==306; platform_system == "Windows"
wmi==1.5.1; platform_system == "Windows"
pyobjc==9.2; platform_system == "Darwin"
psutil==5.9.5
watchdog==3.0.0

# Cloud & Remote Access
firebase-admin==6.2.0
python-telegram-bot==20.6
twilio==8.10.0  # SMS
pusher==3.3.2  # Real-time push
redis==5.0.0
celery==5.3.4  # Async tasks

# Network Monitoring
scapy==2.5.0
netifaces==0.11.0

# Machine Learning & Scanning
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2
yara-python==4.3.1  # Malware pattern matching
clamd==1.0.2  # ClamAV integration

# VirusTotal API
vt-py==0.17.0  # Official VirusTotal API

# Database & Storage
sqlalchemy==2.0.19
pymongo==4.5.0  # MongoDB for cloud sync

# Dashboard & Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Security & Encryption
cryptography==41.0.7
pycryptodome==3.19.0
